rearange and move files
