+++ index.html
@@ -84,7 +84,7 @@
 
 
 
-<script type="module" src="./scripts/mainmain.js"></script>
+<script type="module" src="./scripts/main.js"></script>
 <script src="./scripts/modules/score.js"></script>
 </body>
 </html>
